Apply clang-format

+++ code/matrix.cpp
@@ -1,13 +1,15 @@
 #include "matrix.h"
+
+#include <cassert>
 #include <iostream>
 
-matrix *outer(vector *a, vector *b) {
-	auto *out = new matrix(a->size());
+matrix* outer(vector* a, vector* b) {
+	auto* out = new matrix(a->size());
 
-	for (int i = 0; i < a->size(); i++) {
+	for (size_t i = 0; i < a->size(); i++) {
 		out->at(i) = vector(b->size());
 
-		for (int j = 0; j < b->size(); j++) {
+		for (size_t j = 0; j < b->size(); j++) {
 			out->at(i).at(j) = a->at(i) * b->at(j);
 		}
 	}
@@ -15,15 +17,15 @@ matrix *outer(vector *a, vector *b) {
 	return out;
 }
 
-matrix *add(matrix *a, matrix *b) {
+matrix* add(matrix* a, matrix* b) {
 	assert(a->size() == b->size());
 
-	auto *c = new matrix(a->size());
-	for (int i = 0; i < a->size(); i++) {
+	auto* c = new matrix(a->size());
+	for (size_t i = 0; i < a->size(); i++) {
 		assert(a->at(i).size() == b->at(i).size());
 		c->at(i) = vector(a->at(i).size());
 
-		for (int j = 0; j < a->at(i).size(); j++) {
+		for (size_t j = 0; j < a->at(i).size(); j++) {
 			c->at(i).at(j) = a->at(i).at(j) + b->at(i).at(j);
 		}
 	}
@@ -32,8 +34,8 @@ matrix *add(matrix *a, matrix *b) {
 }
 
 void matrix::print() {
-	for (int i = 0; i < this->size(); i++) {
-		for (int j = 0; j < this->at(i).size(); j++) {
+	for (size_t i = 0; i < this->size(); i++) {
+		for (size_t j = 0; j < this->at(i).size(); j++) {
 			std::cout << this->at(i).at(j) << " ";
 		}
 
@@ -41,25 +43,28 @@ void matrix::print() {
 	}
 }
 
-void matrix::add(matrix *b) {
+void matrix::add(matrix* b) {
 	assert(this->size() == b->size());
 
-	for (int i = 0; i < this->size(); i++) {
+	for (size_t i = 0; i < this->size(); i++) {
 		assert(this->at(i).size() == b->at(i).size());
 
-		for (int j = 0; j < this->at(i).size(); j++) {
+		for (size_t j = 0; j < this->at(i).size(); j++) {
 			this->at(i).at(j) += b->at(i).at(j);
 		}
 	}
 }
 
-bool matrix::equal(matrix *b) {
-	if (this->size() != b->size()) return false;
-	for (int i = 0; i < this->size(); i++) {
-		if (this->at(i).size() != b->at(i).size()) return false;
+bool matrix::equal(matrix* b) {
+	if (this->size() != b->size())
+		return false;
+	for (size_t i = 0; i < this->size(); i++) {
+		if (this->at(i).size() != b->at(i).size())
+			return false;
 
-		for (int j = 0; j < this->at(i).size(); j++) {
-			if (this->at(i).at(j) != b->at(i).at(j)) return false;
+		for (size_t j = 0; j < this->at(i).size(); j++) {
+			if (this->at(i).at(j) != b->at(i).at(j))
+				return false;
 		}
 	}
